#107 Follow template layout yaml
