+++ examples/pySocket/arch/yaml/pySocket.yaml
@@ -1,38 +1,5 @@
 
-types:
-  param_t:
-    width: 32
-    desc: "Parameter type"
-  word16_t:
-    width: 16
-    desc: "Parameter type"
-
-enums:
-  p2s_message_ID_t:
-    desc: "Message ID"
-    enum:
-      - {enumName: P2S_MESSAGE_TYPE_REQUEST, value: 0, desc: "Request"}
-      - {enumName: P2S_MESSAGE_TYPE_RESPONSE, value: 1, desc: "Response"}
-
-
-structures:
-  message_header_st:
-    length: {varType: word16_t, desc: "Message payload length"}
-    ID:     {varType: word16_t, desc: "Message ID"}
-    tag:    {varType: word16_t, desc: "Tag"}
-  p2s_message_st:
-    param2: {varType: param_t, desc: "Parameter 2"}
-    param1: {varType: param_t, desc: "Parameter 1"}
-  p2s_response_st:
-    response: {varType: param_t, desc: "response"}
-
 blocks:
-  pySocket_tb:
-    desc: Module Testbench Top (Root Container)
-    hasVl: false
-    hasRtl: false
-    hasMdl: true
-    hasTb: false
   pySocket:
     desc: Module Top
     hasVl: true
@@ -45,63 +12,3 @@ blocks:
     hasRtl: true
     hasMdl: true
     hasTb: false
-
-instances:
-  pySocket_tb:  { container: pySocket_tb, instanceType: pySocket_tb,     instGroup: top  }
-  u_pySocket:   { container: pySocket_tb, instanceType: pySocket,        instGroup: top  }
-  u_dut:        { container: pySocket_tb, instanceType: dut,             instGroup: top  }
-
-
-
-# interfaces are needed to make any connection from one instance of a block to another block's instance
-## interfacename: is the interface's name and key for usage by any other dictionary doing a lookup, required
-### interfaceType: followed by a string that links to a valid interface type
-####        i.e. rdy_vld could be defined as a ready valid interface
-####             static could be defined as a single port interface with a structure and no protocol
-####             req_ack could be defined as a request acknowledge interface with a structure in both directions
-### structures: is a list of dictionaries, each dictionary has two keys, each item in the list is one per structure on this interface
-###                 some interfaces can have more than one structure but all interfaces must have at least one
-#### dash (-) followed by a dictionary
-#### structureName: followed by a string is a key reference to the structures dictionary
-#### structureType: followed by a string, the structureType is what the structureName gets bound to during generation
-####                    a System Verilog interface may have more than one structureType defined and that is what each 
-####                    structureName gets bound to during generation, similar for System C
-### desc: followed by a string is a description of this interface, required
-interfaces:
-  test_req_ack:
-    interfaceType: req_ack
-    desc: "Req Ack Test interface"
-    structures:
-      - {structure: p2s_message_st, structureType: data_t}
-      - {structure: p2s_response_st, structureType: rdata_t}
-  test2Python_req_ack:
-    interfaceType: req_ack
-    desc: "Req Ack Test2Python interface"
-    structures:
-      - {structure: p2s_message_st, structureType: data_t}
-      - {structure: p2s_response_st, structureType: rdata_t}
-  dut2Python_req_ack:
-    interfaceType: req_ack
-    desc: "Req Ack Dut2Python interface"
-    structures:
-      - {structure: p2s_message_st, structureType: data_t}
-      - {structure: p2s_response_st, structureType: rdata_t}
-
-# connections are specify interface connections between different instaces of objects
-## - {}, dash followed by a dictionary, connections is a list of dictionaries
-### interfacename: followed by string that is a key to this interface, an interface can be connected 1 or more times
-### src: followed by a name of the source / master instance for this conneciton
-### dst: followed by a name of the destination / slave instance for this conneciton
-### count: followed by a number to indicate the number of ports to connect to, only valid with srcport or dstport and an interfacename that has multiple ports
-###
-### ports are optional, and are implied in most use cases. The exception is where disambiguation is needed, mainly in array use cases
-### srcport: used if there is more than one objectname instance of the source / master, declares the port of source objectname to connect to
-### dstport: used if there is more than one objectname instance of destination / slave, declares the port of destination objectname to connect to
-### color: followed by a string which is a color (provide link to valid colors) this adds a color to diagrams
-###          for this connection, optional (currently does not work in diagram generation code)
-connections:
-  - { interface: test_req_ack, src: u_pySocket, dst: u_dut }
-  - { interface: test2Python_req_ack, src: u_pySocket, dst: u_dut }
-  - { interface: dut2Python_req_ack, src: u_dut, dst: u_pySocket }
-
-
