+++ examples/pySocket/fw/include/pySocket_tbIncludesFW.cpp
@@ -1,0 +1,12 @@
+
+// 
+#include "pySocket_tbIncludesFW.h"
+using namespace fw_ns;
+// GENERATED_CODE_PARAM --context=pySocket_tb.yaml --mode=fw
+// GENERATED_CODE_BEGIN --template=structures --section=cppIncludes
+
+// GENERATED_CODE_END
+// GENERATED_CODE_BEGIN --template=structures --section=cpp --namespace=fw_ns
+// structures
+
+// GENERATED_CODE_END
