+++ examples/pySocket/verif/vl_wrap/dut_hdl_sv_wrapper.sv
@@ -6,7 +6,7 @@
 
 module dut_hdl_sv_wrapper
     // Generated Import package statement(s)
-    import pySocket_package::*;
+    import pySocket_tb_package::*;
 (
     // req_ack_if.dst
     input bit test_req_ack_req,
